move and add some more tests
